+++ cogs/helpers.py
@@ -63,51 +63,6 @@ class Helpers(commands.Cog):
     def __init__(self, bot):
         self.bot = bot
 
-    async def cfia_rss(self):
-        # Written by @jidicula
-        """
-        Co-routine that periodically checks the CFIA Health Hazard Alerts RSS
-         feed for updates.
-        """
-        await self.bot.wait_until_ready()
-        while not self.bot.is_closed():
-            recall_channel = utils.get(self.bot.get_guild(
-                self.bot.config.server_id).text_channels,
-                                       name=self.bot.config.recall_channel)
-            newest_recalls = feedparser.parse(CFIA_FEED_URL)['entries']
-            try:
-                id_unpickle = open("pickles/recall_tag.obj", 'rb')
-                recalls = pickle.load(id_unpickle)
-                id_unpickle.close()
-            except Exception:
-                recalls = {}
-            new_recalls = False
-            for recall in newest_recalls:
-                recall_id = recall['id']
-                if recall_id not in recalls:
-                    new_recalls = True
-                    recalls[recall_id] = ""
-                    recall_warning = discord.Embed(title=recall['title'],
-                                                   description=recall['link'])
-                    soup = BeautifulSoup(recall['summary'], "html.parser")
-                    try:
-                        img_url = soup.img['src']
-                        summary = soup.p.find_parent().text.strip()
-                    except Exception:
-                        img_url = ""
-                        summary = recall['summary']
-                    if re.search(self.bot.config.recall_filter, summary,
-                                 re.IGNORECASE):
-                        recall_warning.set_image(url=img_url)
-                        recall_warning.add_field(name="Summary", value=summary)
-                        await recall_channel.send(embed=recall_warning)
-            if new_recalls:
-                # Pickle newly added IDs
-                id_pickle = open("pickles/recall_tag.obj", 'wb')
-                pickle.dump(recalls, id_pickle)
-                id_pickle.close()
-            await asyncio.sleep(12 * 3600)    # run every 12 hours
-
     @commands.command(aliases=['exams'])
     async def exam(self, ctx):
         """Retrieves the exam schedule link from McGill's Exam website."""
